% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{addDirection}
\alias{addDirection}
\title{addDirection}
\usage{
addDirection(df)
}
\arguments{
\item{df}{- a data.frame with the corrected avg_1min_MW}
}
\value{
the input data.frame with the additional variable of the Direction
}
\description{
This method adds an additional variable called Direction to the already preprocessed input data.frame which contains the corrected avg_1min_MW
}
