% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{addTarif}
\alias{addTarif}
\title{addTarif}
\usage{
addTarif(df)
}
\arguments{
\item{df}{- a data.frame which has to be preprocessed and additional information about the tarif is needed. The input data.frame has to have at least a DateTime variable (POSIXct object) with hourly resolution.}
}
\value{
the input data.frame with the additional variable of the tarif
}
\description{
This method adds an additional variable called Tarif to the already preprocessed input data.frame which contains the right DateTime (@seealso preprocessOperatingReserveNeeds or preprocessOperatingReserveCalls). It depends on the helper method @seealso isGermanHoliday
}
