% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{aggregateXminAVGMW}
\alias{aggregateXminAVGMW}
\title{aggregateXminAVGMW}
\usage{
aggregateXminAVGMW(ne, xmin)
}
\arguments{
\item{xmin}{- a timeframe in minutes to build the average values}

\item{df.needs}{- the data.frame containing the operating reserve needs (in MW), a Date (as String) and a Time (as String) variable. As well as type ("RZBedarf")}
}
\value{
data.frame with the Xmin (normally 1min) DateTime (POSIXct object) and the average operating reserve need value of the Xmin time window
}
\description{
This method averages the data.frame of the operating reserve needs on a X minute time frame (e.g. 1 minute) from the input needs (4sec resolution).
}
\examples{
df.needs <- getOperatingReserveNeeds("30.12.2015", "30.12.2015")
aggregateXminAVGMW(df.needs, 1)

}
