% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{approximateCalls}
\alias{approximateCalls}
\title{approximateCalls}
\usage{
approximateCalls(reserveNeeds, reserveCalls)
}
\arguments{
\item{reserveNeeds}{- The preprocessed operating reserve needs (@seealso preprocessOperatingReserveNeeds)}

\item{reserveCalls}{- The preprocessed operating reserve calls (@seealso preprocessOperatingReserveCalls)}
}
\value{
A complete data.frame with all needed variables for correcting the operating reserve needs and approximating the calls.
}
\description{
This method builds the data.frame with all needed variables to correct the operating reserve needs power for approximating the calls.
}
\details{
---------------------------------------------------------

MAIN FUNCTIONS
}
\examples{
df.needs <- getOperatingReserveNeeds("30.12.2015", "30.12.2015")
df.calls <- getOperatingReserveCalls('30.12.2015', '30.12.2015', '6', 'SRL')

reserveCalls <- preProcessOperatingReserveCalls(df.calls)
reserveNeeds <- preprocessOperatingReserveNeeds(df.needs)

approximateCalls(reserveNeeds, reserveCalls)

}
