% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{approximateCallsInRecursion}
\alias{approximateCallsInRecursion}
\title{approximateCalls}
\usage{
approximateCallsInRecursion(df.needs, df.calls)
}
\arguments{
\item{df.needs}{- The preprocessed operating reserve needs (@seealso preprocessOperatingReserveNeeds)}

\item{df.calls}{- The preprocessed operating reserve calls (@seealso preprocessOperatingReserveCalls)}
}
\value{
A complete data.frame with the corrected operating reserve needs to approximate the 1min calls (avg_1min_MW).
}
\description{
This recursive approach handles the special case of CrossingZero. It builds the data.frame with all needed variables to correct the operating reserve needs power for approximating the calls.
}
\examples{
df.needs <- getOperatingReserveNeeds("30.12.2015", "30.12.2015")
df.calls <- getOperatingReserveCalls('30.12.2015', '30.12.2015', '6', 'SRL')

df.calls <- preProcessOperatingReserveCalls(df.calls)
df.needs <- preprocessOperatingReserveNeeds(df.needs)

approximateCallsInRecursion(df.needs, df.calls)


}
