% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{approximateOperatingReserveCalls}
\alias{approximateOperatingReserveCalls}
\title{approximateOperatingReserveCalls}
\usage{
approximateOperatingReserveCalls(reserveNeeds)
}
\arguments{
\item{The}{data.frame of specific 4sec operating reserve needs which is used to approximate the a 1min resolution for the reserve calls.}
}
\value{
A data.frame which approximates the 1min reserve calls
}
\description{
THis function approximates a finer 1min resolution of the 15min operating reserve calls. It does this by modifying the values of the operating reserve needs which are available in a 4sec resolution.
}
\details{
MAIN FUNCTIONS
}
\examples{
approximateOperatingReserveCalls(operatingReserveNeeds)

}

