% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{approxRecursionWith15minChunk}
\alias{approxRecursionWith15minChunk}
\title{This is a helper method used in @seealso approximateCallsInRecursion
It handles the recursion.}
\usage{
approxRecursionWith15minChunk(dataframe)
}
\arguments{
\item{dataframe}{- a dataframe of a 15min section. Only 15 observations}
}
\value{
a 15min data.frame with the corrected/updated values of avg_1min_MW
}
\description{
This is a helper method used in @seealso approximateCallsInRecursion
It handles the recursion.
}
