% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeData.R
\name{build_df_rl_calls_auctions}
\alias{build_df_rl_calls_auctions}
\title{---------------------------------------------------------}
\usage{
build_df_rl_calls_auctions(response_content, fileName)
}
\description{
---------------------------------------------------------
}
