% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{buildCorrectingCallsDF}
\alias{buildCorrectingCallsDF}
\title{buildCorrectingCallsDF}
\usage{
buildCorrectingCallsDF(df.needs, df.calls)
}
\arguments{
\item{df.1min}{- The data.frame with the 1min average operating reserve needs}

\item{df.15min.neg}{- The data.frame with the 15min average negative operating reserve needs}

\item{df.15min.pos}{- The data.frame with the 15min average positive operating reserve needs}

\item{df.15min.calls}{- The data.frame with the 15min operating reserve calls (both negative and positive. Already preprocessed)}
}
\value{
A complete data.frame with all needed variables for correcting the operating reserve needs and approximating the calls.
}
\description{
This method builds up the data.frame with all needed variables to correct the operating reserve needs power for approximating the calls.
}
\examples{
df.needs <- getOperatingReserveNeeds("30.12.2015", "30.12.2015")
df.calls <- getOperatingReserveCalls('30.12.2015', '30.12.2015', '4', 'SRL')

buildCorrectingCallsDF(df.needs, df.calls)

}

