% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{calcHomogenityCorrectness}
\alias{calcHomogenityCorrectness}
\title{calcHomogenityCorrectness}
\usage{
calcHomogenityCorrectness(dataframe)
}
\arguments{
\item{dataframe}{- dataframe with 15min neg and pos calls, 1min avg needs, NEG and POS counts and cuttedTime based on 15min.}
}
\value{
An updated dataframe where the NEG and POS counts in the homogenity cases are modified to 14/1 (from 15/0) and the smallest absolute value is updated to 0.
}
\description{
Takes as input a combined minutely data.frame with 15min calls and 1 min avg of needs as well as the counts of negative and positive needs within 15min. It then updates the 1min avg need and NEG and POS values based on the homogenity.
}
