% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{contourCallProbTimeSeries}
\alias{contourCallProbTimeSeries}
\title{Helper function for the @seealso plotContourCallProb function.
It creates the appropriate data set --> parallel computing}
\usage{
contourCallProbTimeSeries(timeSeriesData, mini, maxi, granularity, timesteps,
  numCores)
}
\description{
Helper function for the @seealso plotContourCallProb function.
It creates the appropriate data set --> parallel computing
}
