% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{correctionCalculationForRecursion}
\alias{correctionCalculationForRecursion}
\title{This helping method is called in the @seealso approxRecursionWith15minChunk function to update the avg_1min_MW variable in a recursively manner.
Here no extra variable Corrected is added!
It takes a data.frame as input with the avg_1min_MW variable and the additional homogenity variables.}
\usage{
correctionCalculationForRecursion(res)
}
\description{
This helping method is called in the @seealso approxRecursionWith15minChunk function to update the avg_1min_MW variable in a recursively manner.
Here no extra variable Corrected is added!
It takes a data.frame as input with the avg_1min_MW variable and the additional homogenity variables.
}
