% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{countCrossingZeros}
\alias{countCrossingZeros}
\title{countCrossingZeros}
\usage{
countCrossingZeros(df)
}
\arguments{
\item{df}{- the input data.frame with a cuttedTime variable of 15min sections. The data.frame has to be preprocces and approximated (@seealso approximateCalls)}
}
\value{
The input data.frame with two additional variables (columns) namely n_negCZ and n_pos_CZ. CZ stands for CrossingZero - the special case.
}
\description{
This function needs the helper methods countNegativeCrossingZerosForDF and countPositiveCrossingZerosForDF. It then adds 2 additional variables to the input data.farme. n_neg_CZ and n_pos_CZ For every 15min section. The numbers in the columns represent the total amount of each CrossingZero type within the 15min section.
}
\examples{
# ... data already crawled and preprocessed
df <- approximateCalls(preprocessed.needs, preprocessed.calls)

df.with.cz.counts <- countCrossingZeros(df)

}
