% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{countNegativeCrossingZerosForDF}
\alias{countNegativeCrossingZerosForDF}
\title{This helper method is used in the countCrossingZeros function.
It counts the corrected needs which were positive and are now negative within the input data.frame.}
\usage{
countNegativeCrossingZerosForDF(df)
}
\description{
This helper method is used in the countCrossingZeros function.
It counts the corrected needs which were positive and are now negative within the input data.frame.
}
