% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{countPositiveCrossingZerosForDF}
\alias{countPositiveCrossingZerosForDF}
\title{This helper method is used in the countCrossingZeros function.
It counts the corrected needs which were negative and are now positive within the input data.frame.}
\usage{
countPositiveCrossingZerosForDF(df)
}
\description{
This helper method is used in the countCrossingZeros function.
It counts the corrected needs which were negative and are now positive within the input data.frame.
}
