% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{get15minAVGOff1minAVG}
\alias{get15minAVGOff1minAVG}
\title{get15minAVGOff1minAVG}
\usage{
get15minAVGOff1minAVG(dataframe, direction)
}
\arguments{
\item{dataframe}{- the data.frame with operating reserve needs (already preprocessed)}

\item{direction}{- specifies the positive or negative operating reserve need}

\item{xmin}{- the time window for which the average should be calculated (e.g. 15)}
}
\value{
a data.frame with corresponding average values on the given time window ("cuttedTime")
}
\description{
This method takes a data.frame of 1 min average values of operating reserve needs (@seealso aggregateXminAVGMW) and calculates the average values of a specific time window.
}
\examples{
df.needs <- getOperatingReserveNeeds("30.12.2015", "30.12.2015")
df.needs.preprocessed <- preprocessOperatingReserveNeeds(df.needs)
df.avg.15min.neg <- getAVGXmin(df.needs.preprocessed, 15, "NEG")

}

