% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{get15minAVGs}
\alias{get15minAVGs}
\title{This helper method is used in @seealso approxRecursionWith15minChunk
It retrieves a data.frame and returns the same data.frame with added columns for NEG and POS 15min avg of the 1min needs}
\usage{
get15minAVGs(dataframe)
}
\description{
This helper method is used in @seealso approxRecursionWith15minChunk
It retrieves a data.frame and returns the same data.frame with added columns for NEG and POS 15min avg of the 1min needs
}
