% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{getAVGXmin}
\alias{getAVGXmin}
\title{getAVGXmin}
\usage{
getAVGXmin(dataframe, xmin, direction)
}
\arguments{
\item{dataframe}{- the data.frame with operating reserve needs (already preprocessed)}

\item{xmin}{- the time window for which the average should be calculated (e.g. 15)}

\item{direction}{- specifies the positive or negative operating reserve need}
}
\value{
a data.frame with corresponding average values on the given time window ("cuttedTime")
}
\description{
This method takes a data.frame (operating reserve needs) and gets the corresponding average values of a specific time window.
}
\examples{
df.needs <- getOperatingReserveNeeds("30.12.2015", "30.12.2015")
df.needs.preprocessed <- preprocessOperatingReserveNeeds(df.needs)
df.avg.15min.neg <- getAVGXmin(df.needs.preprocessed, 15, "NEG")

}

