% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{getCallProbForMarginalWorkPrice}
\alias{getCallProbForMarginalWorkPrice}
\title{This is a helper method needed in the @seealso getCallProbDataSet function
It retuns the call probability for a given price (mwp) based on the conditions}
\usage{
getCallProbForMarginalWorkPrice(data, mwp, conditionByColumns)
}
\description{
This is a helper method needed in the @seealso getCallProbDataSet function
It retuns the call probability for a given price (mwp) based on the conditions
}
