% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{getHighestPriceWithHighestCallProb}
\alias{getHighestPriceWithHighestCallProb}
\title{getHighestPriceWithHighestCallProb}
\usage{
getHighestPriceWithHighestCallProb(data, startPrice, endPrice, direction,
  granularityPrice, granularityTime, numCores)
}
\arguments{
\item{data}{- data.frame}

\item{startPrice}{-}

\item{endPrice}{-}

\item{direction}{- filters the plot by the direction of the reserve energy. String as "POS" or "NEG"}

\item{granularityPrice}{-}

\item{granularityTime}{-}

\item{numCores}{-}
}
\value{
data.frame with highest probabilities in Prob column and the corresponding highest price in the Price column for each time step
}
\description{
This method creates a data.frame with same variables as the input data.frame but filter for the highest call probabilities with highest price at each time step.
}
\examples{
marginalworkprices <- getMarginalWorkPrices(needs, calls, auctions)
# Get the data.frame with highest call probability and its highest work price for each time step.
# price range from 0 to 80 with whole steps (0,1,2,...,79,80) and hourly time steps (60 minutes)
getHighestPriceWithHighestCallProb(m, 0, 80, "POS", 1, 60, 2)

}
