% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{getMarginalWorkPrices}
\alias{getMarginalWorkPrices}
\title{getMarginalWorkPrice}
\usage{
getMarginalWorkPrices(df, auctions)
}
\arguments{
\item{df}{- The preprocessed and approximated calls (@seealso approximateCallsInRecursion)}

\item{auctions}{- The preprocessed operating reserve auctions (@seealso getOperatingReserveAuctions and preprocessOperatingReserveAuctions)}
}
\value{
A complete data.frame with the 1min approximated calls and the corresponding marginal work prices for every minute
}
\description{
This recursive approach handles the special case of CrossingZero. It builds the data.frame with all needed variables to correct the operating reserve needs power for approximating the calls.
}
\examples{
sample.auctions <- getOperatingReserveAuctions('28.12.2015', '07.01.2016', '2')
sample.needs <- getOperatingReserveNeeds("01.01.2016", "01.01.2016")
sample.calls <- getOperatingReserveCalls('01.01.2016', '01.01.2016', '6', 'SRL')

s.c <- preProcessOperatingReserveCalls(sample.calls)
s.n <- preprocessOperatingReserveNeeds(sample.needs)
s.a <- preprocessOperatingReserveAuctions(sample.auctions)

df.approx.calls <- approximateCallsInRecursion(s.n, s.c)

df <- getMarginalWorkPrices(df.approx.calls, s.a)


}
