% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{getNumberOfPosOrNegIn15min}
\alias{getNumberOfPosOrNegIn15min}
\title{getNumberOfPosOrNegIn15min}
\usage{
getNumberOfPosOrNegIn15min(dataframe)
}
\arguments{
\item{dataframe}{- THe input data.frame has to have already a cuttedTime variable with 15min sections! (trying to minimize the amount of cut operations due to its computational intensivity)}
}
\value{
a data.frame with corresponding counting values for NEG and POS and adds additionally a Direction variable (NEG, POS)
}
\description{
This method adds the counts of NEG and POS 1min averages for every 15min section.
}
\examples{
No example! This function is included in the buildCorrectingCallsDF function.

}
