% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minScript.R
\name{getOneMinuteCalls}
\alias{getOneMinuteCalls}
\title{getOneMinuteCalls}
\usage{
getOneMinuteCalls(needs, calls, numCores)
}
\arguments{
\item{needs}{- The preprocessed operating reserve needs (@seealso getReserveNeeds)}

\item{calls}{- The preprocessed operating reserve calls (@seealso getReserveCalls)}

\item{numCores}{- OPTIONAL PARAMETER --> PARALLEL COMPUNTING NOT YET FUNCTIONAL}
}
\value{
A complete data.frame with the corrected operating reserve needs to approximate the 1min calls (avg_1min_MW).
}
\description{
This production function calls the @seealso approximateCallsInRecursion() method in @seealso preprocessData.R script.
It calculaates the approximated 1 min calls out of the 4sec reserve needs.
Variables in the returned data.frame:  - neg_MW (numeric). The signed power calls within the 15min time window. Point as decimal delimiter
                                       - pos_MW (numeric). The signed power calls within the 15min time window. Point as decimal delimiter
                                       - approx_1min_call (numeric). The signed, approximated 1min calls. Point as decimal delimiter
                                       - DateTime (POSIXct). DateTime object in Y-m-d h:m:s format. Be aware of daylight saving. 1min time windows.
}
\examples{

needs <- getReserveNeeds('07.03.2017', '14.03.2017')
calls <- getReserveCalls('07.03.2017', '14.03.2017', '6', 'SRL')

OneMinuteCalls <- get1minCalls(needs, calls)


}
