% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeData.R
\name{getOperatingReserveCalls}
\alias{getOperatingReserveCalls}
\title{getOperatingReserveCalls}
\usage{
getOperatingReserveCalls(date_from, date_to, uenb_type, rl_type)
}
\arguments{
\item{date_from}{sets the starting date in format: DD.MM.YYYY}

\item{date_to}{sets the ending date in format: DD.MM.YYYY}

\item{uenb_type}{[50Hz (4), TenneT (2), Amprion (3), TransnetBW (1), Netzregelverbund (6), IGCC (11)]}

\item{rl_type}{[SRL, MRL, RZ_SALDO, REBAP, ZUSATZMASSNAHMEN, NOTHILFE]}
}
\value{
data.frame variable containing the operating reserve call table
}
\description{
This main function retrieves the operating reserve calls from \url{https://www.regelleistung.net/ext/data/}. The time interval can only lay within the same month.
}
\examples{
getOperatingReserveCalls('07.03.2017', '14.03.2017', '4', 'SRL')

}

