% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeData.R
\name{getOperatingReserveNeeds}
\alias{getOperatingReserveNeeds}
\title{getOperatingReserveNeeds}
\usage{
getOperatingReserveNeeds(startDate, endDate)
}
\arguments{
\item{date_from}{sets the starting date in format: DD.MM.YYYY}

\item{date_to}{sets the ending date in format: DD.MM.YYYY}
}
\value{
data.frame variable containing the operating reserve need table for the specified time period
}
\description{
This main function retrieves the operating reserve needs from \url{https://www.transnetbw.de/de/strommarkt/systemdienstleistungen/regelenergie-bedarf-und-abruf}. The resolution is 4sec. The function can take awhile since it has to download sever MBs of data. The oldest data that can be retrieved is July (07) 2010.
}
\examples{
getOperatingReserveNeeds("30.12.2015", "02.01.2016")

}

