% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minScript.R
\name{getReserveAuctions}
\alias{getReserveAuctions}
\title{getReserveAuctions}
\usage{
getReserveAuctions(startDate, endDate, rl)
}
\arguments{
\item{startDate}{the start date to retrieve all auctions. Format (german style): DD.MM.YYYY (e.g.'07.03.2017')}

\item{endDate}{the end date to retrieve all auctions. Format (german style): DD.MM.YYYY (e.g.'07.03.2017')}

\item{rl}{PRL (1), SRL (2), MRL (3), sofort abschaltbare Lasten (4), schnell abschaltbare Lasten (5), Primärregelleistung NL (6)}
}
\value{
data.frame with the results of the auctions held from start date till end date
}
\description{
This function is the production method of @seealso getOperatingReserveAuctions in the @seealso scrapeData.R script. It is without console prints for faster computation and several steps are combined.
This main function retrieves the operating reserve auction results from \url{https://www.regelleistung.net/ext/tender/}.
The data contains all auctions from a given start date till an end date. Be aware of the weekly data and take care of the latest week.
Variables in the returned data.frame:  - power_price (numeric). The signed power price of the bid. Point as decimal delimiter.
                                       - work_price (numeric). The signed work price of the bid. Point as decimal delimiter. There are also negative prices.
                                       - offered_power_MW (numeric). The offered power in MW. Point as decimal delimiter. 5 as the minimal bid. This will be changed someday to smaller increments.
                                       - Tarif (char). Contains the value of either "HT", the main tarif or "NT", for the remaining time periods.
                                       - Direction (char). Contains the value of either "POS", for positive operating reserve power, or "NEG" for negative power.
                                       - date_from (Date). Date object in Y-m-d format. The two date variables build the one week auction time period. This will be changed someday.
                                       - date_to (Date). Date object in Y-m-d format. The two date variables build the one week auction time period. This will be changed someday.
}
\examples{
# Get the auction results of the secondary operating reserve power for one week of that specific day
auctions <- getReserveAuctions('07.03.2017','07.03.2017', '2')

}
