% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minScript.R
\name{getReserveCalls}
\alias{getReserveCalls}
\title{getReserveCalls}
\usage{
getReserveCalls(startDate, endDate, uenb, rl)
}
\arguments{
\item{startDate}{- sets the starting date. Format (german style): DD.MM.YYYY}

\item{endDate}{- sets the ending date. Format (german style): DD.MM.YYYY}

\item{uenb}{- 50Hz (4), TenneT (2), Amprion (3), TransnetBW (1), Netzregelverbund (6), IGCC (11)}

\item{rl}{- SRL, MRL, RZ_SALDO, REBAP, ZUSATZMASSNAHMEN, NOTHILFE}
}
\value{
data.frame variable containing the operating reserve call table
}
\description{
This function is the production method of @seealso getOperatingReserveCalls in the @seealso scrapeData.R script. It is without console prints for faster computation and several steps are combined.
This main function retrieves the 15min operating reserve calls (not qualified) from \url{https://www.regelleistung.net/ext/data/}.
The oldest data that can be retrieved is 2011-06-27.
The method also preprocesses the data in a nice format.
Variables in the returned data.frame:  - neg_MW (numeric). The signed power calls within the 15min time window. Point as decimal delimiter
                                       - pos_MW (numeric). The signed power calls within the 15min time window. Point as decimal delimiter
                                       - DateTime (POSIXct). DateTime object in Y-m-d h:m:s format. Be aware of daylight saving. 15min time windows.
}
\examples{
# Get the secondary calls for one week of the Netzregelverbund
calls <- getReserveCalls('07.03.2017', '14.03.2017', '6', 'SRL')

}
