% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minScript.R
\name{getReserveNeeds}
\alias{getReserveNeeds}
\title{getReserveNeeds}
\usage{
getReserveNeeds(startDate, endDate)
}
\arguments{
\item{startDate}{sets the starting date. Format (german style): DD.MM.YYYY}

\item{endDate}{sets the ending date. Format (german style): DD.MM.YYYY}
}
\value{
data.frame variable containing the 4sec operating reserve needs.
}
\description{
This function is the production method of @seealso getOperatingReserveNeeds in the @seealso scrapeData.R script. It is without console prints for faster computation and several steps are combined.
This main function retrieves the operating reserve needs from \url{https://www.transnetbw.de/de/strommarkt/systemdienstleistungen/regelenergie-bedarf-und-abruf}.
The resolution is 4sec. The function can take awhile since it has to download sever MBs of data. The oldest data that can be retrieved is July (07) 2010.
The method also preprocesses the data in a nice format.
Variables in the returned data.frame:  - MW (numeric). The signed power needs for every 4sec. Point as decimal delimiter
                                       - DateTime (POSIXct). DateTime object in Y-m-d h:m:s format. Be aware of daylight saving. 4sec time windows.
}
\examples{
needs <- getReserveNeeds("30.12.2015", "02.01.2016")

}
