% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{getTimeSeriesProbForPrice}
\alias{getTimeSeriesProbForPrice}
\title{Helper function fot the @seealso plotMWPTimeSeries function.}
\usage{
getTimeSeriesProbForPrice(data, price, conditionOnTarifAndDirection,
  granularity)
}
\description{
It produces the right data set which can than be plotted.
condition is only based on Tarif and Direction
granularity of time in minutes. E.g. 60 --> 60min leads to hourly data
}
