% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{imputeMissingCallsWithTSO}
\alias{imputeMissingCallsWithTSO}
\title{imputeMissingCallsWithTSO}
\usage{
imputeMissingCallsWithTSO(calls, tso.list)
}
\arguments{
\item{calls}{- the call data.frame with missing values which has to be imputed}

\item{tso.list}{- list of the call data of the 4 tso companies (data must be same time period as call data).}
}
\value{
the imputed call data.frame
}
\description{
This method replaces missing values for 15 minute calls. It uses the data of the 4 TSOs. The sum equals the values of the Netzregelverbund. If also for the TSO data is missing than it uses the data of the day before.
NOTE: If there is a missing negative value there is also a missing positive value (Checked for 2013,2012 and 2011)
}
\examples{
# Get call data with missing values
calls = getReserveCalls('01.01.2013', '31.12.2013', '6', 'SRL')

# 50Hz (4)
calls.2013.4 = getReserveCalls('01.01.2013', '31.12.2013', '4', 'SRL')
# TenneT (2)
calls.2013.2 = getReserveCalls('01.01.2013', '31.12.2013', '2', 'SRL')
# Amprion (3)
calls.2013.3 = getReserveCalls('01.01.2013', '31.12.2013', '3', 'SRL')
# TransnetBW (1)
calls.2013.1 = getReserveCalls('01.01.2013', '31.12.2013', '1', 'SRL')
# Build the tso list
tso.list <- list(calls.2013.1,calls.2013.2,calls.2013.3,calls.2013.4)

# Impute the data
imputed.calls <- imputeMissingCallsWithTSO(calls, tso.list)

}
