% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{markCrossingZero}
\alias{markCrossingZero}
\title{markCrossingZero}
\usage{
markCrossingZero(df)
}
\arguments{
\item{df}{- the input data.frame with a cuttedTime variable of 15min sections. The data.frame has to be preprocces and approximated (@seealso approximateCalls)}
}
\value{
The input data.frame with two additional variables (columns) namely pos_CZ and neg_CZ. CZ stands for CrossingZero - the special case.
}
\description{
This function adds 2 variables to the input data.frame. pos_CZ and neg_CZ. They have the value 1 if the corrected need crossed the zero line. The artificial zero values in the homogenity case are excluded.
}
\examples{
# ... data already crawled and preprocessed
df <- approximateCalls(preprocessed.needs, preprocessed.calls)

df.with.marked.cz <- markCrossingZero(df)

}
