% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeData.R
\name{mergeCSVFilesOfNeeds}
\alias{mergeCSVFilesOfNeeds}
\title{mergeCSVFilesOfNeeds}
\usage{
mergeCSVFilesOfNeeds(year)
}
\arguments{
\item{year}{- the year where all 12 monthly data are bind together into one data.frame}
}
\value{
A data.frame of the merged 12 month operating reserve needs (Date, Time, Type, MW)
}
\description{
This method merges the monthly operating reserve needs which had been downloaded and unzip from the site https://www.transnetbw.de/de/strommarkt/systemdienstleistungen/
}
\examples{
# The 12 month data has already been unziped and stored into ../data/needs
df.needs.2016 <- mergeCSVFilesOfNeeds(2016)

}
