% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{plotApproximatedData}
\alias{plotApproximatedData}
\title{plotApproximatedData}
\usage{
plotApproximatedData(dataframe)
}
\arguments{
\item{dataframe}{- data.frame with the averages (15min of neg. and pos.) of the operating reserve needs and the 15min calls and the approximated 1min calls (avg_1min_MW).}
}
\value{
a line chart
}
\description{
This method plots the 1min average of the operating reserve needs, 15min negative and positive needs, as well as the 15min
}
\examples{
dataframe <- approximateCallsInRecursion(s.needs, s.calls)
plotApproximatedData(dataframe)

}
