% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{plotAVGMWvs4secMW}
\alias{plotAVGMWvs4secMW}
\title{plotXminAVGMWvs4secMW}
\usage{
plotAVGMWvs4secMW(dataframe)
}
\arguments{
\item{dataframe}{- data.frame with the operating reserve needs (in MW), the type ("RZBedarf"), a DateTime (POSIXct object) and the averages (1min, 15min neg. and pos.) operating reserve needs and the 15min calls.}
}
\value{
a line chart with the 4sec operating reserve needs (grey)...
}
\description{
This method plots the 1min average of the operating reserve needs, 15min negative and positive needs, as well as the 15min calls against the original 4sec operating reserve needs
}
\details{
The plotData script
}
\examples{
dataframe <- buildCorrectingCallsDF(df.1min, df.avg.15min.neg, df.avg.15min.pos, df.calls.preprocessed)
plotXminAVGMWvs4secMW(dataframe)

}
\author{
Timo Wagner, \email{wagnertimo@gmx.de}

It contains main functions to provide several plots for analyzing operating reserve data.

Some useful keyboard shortcuts for package authoring:

 Build and Reload Package:  'Cmd + Shift + B'
 Check Package:             'Cmd + Shift + E'
 Test Package:              'Cmd + Shift + T'
}

