% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{plotContourCallProb}
\alias{plotContourCallProb}
\title{plotContourCallProb}
\usage{
plotContourCallProb(data, startPrice, endPrice, direction, granularityPrice,
  granularityTime, numCores)
}
\arguments{
\item{data}{- data.frame}

\item{startPrice}{- specifies the start price for the price range}

\item{endPrice}{- specifies the end price for the price range}

\item{direction}{- filters the plot by the direction of the reserve energy. String as "POS" or "NEG"}

\item{granularityPrice}{- specifies the price steps within the defined price range (startPrice - endPrice). E.g. 0.1 leads to 0.1, 0.2, 0.3 etc.}

\item{granularityTime}{- specifies the time steps in minutes over the time period defined by its input data.frame. E.g. 60 leads to hourly time series data.}

\item{numCores}{- specifies the number of cores used for the parallel computing operation.}
}
\value{
a contour plot with the call probabilities (y axis) for different work prices (color scheme) specified by the price range over a time series.
}
\description{
This method creates a contour plot (3 dimensions) on the call probability (y axis) and work prices (color scheme) over time.
}
\examples{
marginalworkprices <- getMarginalWorkPrices(needs, calls, auctions)
# Contour plot for call probabilities of work prices (positive reserve energy) between 0 - 80 in whole steps (0,1,2,...,79,80) in an hourly time series data.
plotContourCallProb(m, 0, 80, "POS", 1, 60, 2)

}
