% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{plotContourHourOfDay}
\alias{plotContourHourOfDay}
\title{plotContourHourOfDay}
\usage{
plotContourHourOfDay(data, direction, granularityTime, heatMapOrContour)
}
\arguments{
\item{data}{- data.frame}

\item{direction}{- filters the plot by the direction of the reserve energy. String as "POS" or "NEG"}

\item{granularityTime}{- specifies the time steps in minutes over the time period defined by its input data.frame. E.g. 60 leads to hourly time series data.}

\item{heatMapOrContour}{- input string "heatmap" declares if a ggplot2 heatmap should be used and "contour" sets a plotly contour plot as the plot function.}
}
\value{
plot object
}
\description{
This method plots the average marginal work prices of the input data (@seealso getMarginalWorkPrices) in a color scheme against the daily x axis and hourly y axis
}
\examples{
marginalworkprices <- getMarginalWorkPrices(needs, calls, auctions)
# Get the data.frame with highest call probability and its highest work price for each time step.
# price range from 0 to 80 with whole steps (0,1,2,...,79,80) and hourly time steps (60 minutes)
getHighestPriceWithHighestCallProb(m, 0, 80, "POS", 1, 60, 2)

}
