% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{plotCorrectedNeeds}
\alias{plotCorrectedNeeds}
\title{plotCorrectedNeeds}
\usage{
plotCorrectedNeeds(dataframe)
}
\arguments{
\item{dataframe}{- data.frame with the operating reserve needs (in MW), and the averages (1min, 15min neg. and pos.) operating reserve needs and the 15min calls and the corrected needs.}
}
\value{
a line chart with the averages of operating reserve needs (grey)...
}
\description{
This method plots the 1min average of the operating reserve needs, 15min negative and positive needs, as well as the 15min calls against the original 4sec operating reserve needs
}
\examples{
dataframe <- approxOperatingReserveCalls(s.needs, s.calls)
plotCorrectedNeeds(dataframe)

}
