% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{plotMWPTimeSeries}
\alias{plotMWPTimeSeries}
\title{plotMWPTimeSeries}
\usage{
plotMWPTimeSeries(data, price, direction, granularity)
}
\arguments{
\item{data}{- data.frame of the marginal work prices @seealso getMarginalWorkPrices()}

\item{price}{- define a price}

\item{direction}{- filters the plot by the direction of the reserve energy. String as "POS" or "NEG"}

\item{granularity}{- defines the time steps in minutes on the x axis.}
}
\value{
a line chart with time on x axis (depending on the granularity) and call probability on y axis
}
\description{
This method plots the call probabilites over time (specified by the time range of the input data set) for a given price.
The time steps can be defined in minutes. E.g. call probabilities for ever 60 minutes (hourly time series data).
The sign of the reserve energy has also to be specified as an input parameter. This is neccessary to distinguish between the product types.
}
\details{
------------------------------------------------------------
Plotting techniques of marginal work prices, call probabilites etc.
}
\examples{
marginalworkprices <- getMarginalWorkPrices(needs, calls, auctions)
# Plot the call probability time series for price 55 in a 60min interval over the whole time period (range) in the input data set
plotMWPTimeSeries(marginalworkprices, 55, "POS", 60)

}
