% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeData.R
\name{preprocess_rl_calls}
\alias{preprocess_rl_calls}
\title{This function handles the firstly needed preprocessing in the crawling step. It just formats the POST response of @seealso scrape_rl_calls.
Further preprocessing is taken by the functions in the @seealso preprocessData.R script.}
\usage{
preprocess_rl_calls(response_content)
}
\description{
This function handles the firstly needed preprocessing in the crawling step. It just formats the POST response of @seealso scrape_rl_calls.
Further preprocessing is taken by the functions in the @seealso preprocessData.R script.
}
