% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{preprocessOperatingReserveAuctions}
\alias{preprocessOperatingReserveAuctions}
\title{preprocessOperatingReserveAuctions}
\usage{
preprocessOperatingReserveAuctions(df.auctions)
}
\arguments{
\item{df.auctions}{- the data.frame containing the original retrieved operating reserve auctions}
}
\value{
a modified data.frame containing the Type ("RZBedarf"), DateTime (as POSIXct object), the MW and the direction variable (POS or NEG)
}
\description{
This method transforms the original retrieved operating reserve auctions into a nicer data.frame.
}
