% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{preprocessOperatingReserveNeeds}
\alias{preprocessOperatingReserveNeeds}
\title{preprocessOperatingReserveNeeds}
\usage{
preprocessOperatingReserveNeeds(df.needs)
}
\arguments{
\item{df.needs}{- the data.frame containing the original retrieved operating reserve needs}
}
\value{
a modified data.frame containing the Type ("RZBedarf"), DateTime (as POSIXct object), the MW and the direction variable (POS or NEG)
}
\description{
This method transforms the original retrieved operating reserve needs into a nicer data.frame.
}

