% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeData.R
\name{scrape_rl_calls}
\alias{scrape_rl_calls}
\title{The scrapeData script}
\usage{
scrape_rl_calls(date_from, date_to, uenb_type, rl_type)
}
\description{
The scrapeData script
}
\author{
Timo Wagner, \email{wagnertimo@gmx.de}

It contains main and helper functions to crawl call and auction results data out of
}
\references{
\url{https://www.regelleistung.net/ext/data/}

\url{https://www.regelleistung.net/ext/tender/}

Some useful keyboard shortcuts for package authoring:

 Build and Reload Package:  'Cmd + Shift + B'
 Check Package:             'Cmd + Shift + E'
 Test Package:              'Cmd + Shift + T'
}

