% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeData.R
\name{scrape_rl_need_month}
\alias{scrape_rl_need_month}
\title{---------------------------------------------------------}
\usage{
scrape_rl_need_month(date_code)
}
\description{
---------------------------------------------------------
}
