% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minScript.R
\name{setLogging}
\alias{setLogging}
\title{setLogging}
\usage{
setLogging(logger)
}
\arguments{
\item{logger}{- A boolean variable. TRUE for printing out the logging outputs in the console.}
}
\description{
This function sets a global options variable called "logging" to TRUE OR FALSE. By default it is FALSE, so no logging is displayed.
}
\details{
The minScript
}
\references{
@seealso scrapeData.R

@seealso preprocessData.R

Some useful keyboard shortcuts for package authoring:

 Build and Reload Package:  'Cmd + Shift + B'
 Check Package:             'Cmd + Shift + E'
 Test Package:              'Cmd + Shift + T'
}
\author{
Timo Wagner, \email{wagnertimo@gmx.de}

This sscript file contains a minimal version with all necessary functions to crawl, preprocess, callculate the approximated 1min operating reserve calls and calculate the marginal workprices.
All printings in the console are omitted.

TODO: Performance testings: --> parallel computing and maybe substitution of for loops with lapply() functions
}
