% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{aggregateXminAVGMW}
\alias{aggregateXminAVGMW}
\title{aggregateXminAVGMW}
\usage{
aggregateXminAVGMW(df.needs, xmin)
}
\arguments{
\item{df.needs}{- the data.frame containing the operating reserve needs (in MW), a Date (as String) and a Time (as String) variable. As well as type ("RZBedarf")}

\item{xmin}{- a timeframe in minutes to build the average values}
}
\value{
data.frame with the operating reserve needs (in MW), the type ("RZBedarf"), a DateTime (POSIXct object) and the average operating reserve need value of the x min time window
}
\description{
This method averages the data.frame of the operating reserve needs on a x minute time frame (e.g. 1 minute)
}
\examples{
df.needs <- getOperatingReserveNeeds("30.12.2015", "30.12.2015")
aggregateXminAVGMW(df.needs, 1)

}

