% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{get15minAVGs}
\alias{get15minAVGs}
\title{get15minAVGs}
\usage{
get15minAVGs(dataframe)
}
\arguments{
\item{dataframe}{- the data.frame with with operating reserve needs (already preprocessed) and cuttedTime variable 15min is passed in (save computation time!)}
}
\value{
the input data.frame with added columns for 15min neg and pos averages
}
\description{
This helper method is used in @seealso approxRecursionWith15minChunk. It retrieves a data.frame and returns the same data.frame with added columns for NEG and POS 15min avg of the 1min needs
}
\examples{
df.needs <- getOperatingReserveNeeds("30.12.2015", "30.12.2015")
df.needs.preprocessed <- preprocessOperatingReserveNeeds(df.needs)
df.avg.15min.neg <- getAVGXmin(df.needs.preprocessed, "NEG")

}
