% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{getCallProbDataSet}
\alias{getCallProbDataSet}
\title{getCallProbDataSet}
\usage{
getCallProbDataSet(data, numCores, price.seq.start, price.seq.end, tarif,
  direction)
}
\arguments{
\item{data}{- the data.frame with the calculated marginal workprices from the approximated 1min calls (@seealso getMarginalWorkPrices)}

\item{numCores}{- set the number of cores to be used for parallel computation}

\item{price.seq.start}{- specifies the start price of a price range for which call probabilities should be calculated}

\item{price.seq.end}{- specifies the end price of a price range for which call probabilities should be calculated}

\item{tarif}{- name the Tarif ("HT" or "NT") for which call probabilities should be calculated. In combination with the Direction variable this specifies the product type.}

\item{direction}{- name the Direction ("POS" or "NEG") for which call probabilities should be calculated. In combination with the Tarif variable this specifies the product type.}
}
\value{
an array with all the call probabilities of the specified timeperiod and price range.
}
\description{
This functions calculates for each price within a given price range the call probability specified on a time period and product type (Tarif and Direction).
Be aware that undefined time periods and/or Tarifs and Directions can lead to and error/exception which is not handled. So be sure that the time period and Tarif and Direction is valid/exist within the given input data.frame.
}
\examples{
needs <- getOperatingReserveNeeds('01.01.2016', '10.01.2016')
calls <- getOperatingReserveCalls('01.01.2016', '10.01.2016', '6', 'SRL')
auctions <- getOperatingReserveAuctions('28.12.2015', '10.01.2016', '2')

mwork.parallel <- getMarginalWorkPrices(needs,calls,auctions,2)
call.probs <- getCallProbDataSet(mwork.parallel, 1, 0, 775, "2016-01-01 00:00:00", "2016-01-01 23:59:59", "NT", "POS")

# Plot the call probabilities. Therefore create an array with the price range
library(ggplot2)
price.range <- seq(0, ceiling(max.mwork))
qplot(price.range, call.probs, geom="line")


}
