% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{getCallProbForMarginalWorkPrice}
\alias{getCallProbForMarginalWorkPrice}
\title{This is a helper method needed in the @seealso getCallProbDataSet function}
\usage{
getCallProbForMarginalWorkPrice(data, mwp, startTime, endTime, tarif, direction)
}
\description{
This is a helper method needed in the @seealso getCallProbDataSet function
}
