% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minScript.R
\name{getMarginalWorkPrices}
\alias{getMarginalWorkPrices}
\title{getMarginalWorkPrice}
\usage{
getMarginalWorkPrices(needs, calls, auctions, numCores)
}
\arguments{
\item{needs}{- The preprocessed operating reserve needs (@seealso getReserveNeeds)}

\item{calls}{- The preprocessed operating reserve calss (@seealso getReserveCalls)}

\item{auctions}{- The preprocessed operating reserve auctions (@seealso getReserveAuctions)}

\item{numCores}{- This is an optional parameter. If set then the parallelWrapper is used. Specify the number of processor cores. Do not exceed your number of Cores.}
}
\value{
A complete data.frame with the 1min approximated calls and the corresponding marginal work prices for every minute
}
\description{
This production function directly calculates the marginal work prices for every minute (based on the approximated 1min calls).
If you already should have calculated the 1min approximated calls then use the @seealso calcMarginalWorkPrices in the @seealso preprocessData.R file.
}
\examples{

needs <- getReserveNeeds('07.03.2017', '14.03.2017')
calls <- getReserveCalls('07.03.2017', '14.03.2017', '6', 'SRL')
# The time period of the auctions must lie within the period of the calls and needs. An auction week goes from monday till sunday.
auctions <- getReserveAuctions('06.03.2017', '19.03.2017', '2')

df <- getMarginalWorkPrices(needs, calls, auctions)


}
