% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minScript.R
\name{getMarginalWorkPrices}
\alias{getMarginalWorkPrices}
\title{getMarginalWorkPrice}
\usage{
getMarginalWorkPrices(needs, calls, auctions)
}
\arguments{
\item{needs}{- The preprocessed and approximated calls (@seealso getReserveNeeds)}

\item{calls}{- The preprocessed operating reserve auctions (@seealso getReserveCalls)}

\item{auctions}{- The preprocessed operating reserve auctions (@seealso getReserveAuctions)}
}
\value{
A complete data.frame with the 1min approximated calls and the corresponding marginal work prices for every minute
}
\description{
This production function directly calculates the marginal work prices for every minute (based on the approximated 1min calls).
If you already should have calculated the 1min approximated calls then use the @seealso calcMarginalWorkPrices in the @seealso preprocessData.R file.
}
\examples{

needs <- getReserveNeeds('07.03.2017', '14.03.2017')
calls <- getReserveCalls('07.03.2017', '14.03.2017', '6', 'SRL')

OneMinuteCalls <- get1minCalls(needs, calls)

df <- getMarginalWorkPrices(df.approx.calls, s.a)


}
