% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeData.R
\name{getOperatingReserveAuctions}
\alias{getOperatingReserveAuctions}
\title{getOperatingReserveAuctions}
\usage{
getOperatingReserveAuctions(date_from, productId)
}
\arguments{
\item{date_from}{the starting date to retrieve all auctions till NOW in the date format DD.MM.YYYY (e.g.'07.03.2017')}

\item{productId}{PRL (1), SRL (2), MRL (3), sofort abschaltbare Lasten (4), schnell abschaltbare Lasten (5), Primärregelleistung NL (6)}
}
\value{
data.frame with the results of the auctions held from starting date till now
}
\description{
This main function retrieves the operating reserve auction results from \url{https://www.regelleistung.net/ext/tender/}.
The data contains all auctions from a given starting date till now.
}
\examples{
getOperatingReserveAuctions('07.03.2017', '2')

}

