% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{isGermanHoliday}
\alias{isGermanHoliday}
\title{isGermanHoliday}
\usage{
isGermanHoliday(dateTime)
}
\arguments{
\item{dateTime}{- A dateTime object}
}
\value{
a boolean value TRUE if the given dateTime is a german holiday or not (FALSE)
}
\description{
This method states if the given datetime (POSIXct object) is a german bank holiday. It is used e.g. in @seealso addTarif
}
\details{
The preprocessData script
}
\author{
Timo Wagner, \email{wagnertimo@gmx.de}
}
