% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{preProcessOperatingReserveCalls}
\alias{preProcessOperatingReserveCalls}
\title{preProcessOperatingReserveCalls}
\usage{
preProcessOperatingReserveCalls(df.calls)
}
\arguments{
\item{df.calls}{- the data.frame containing the original retrieved operating reserve calls}
}
\value{
a modified data.frame containing the DateTime (as POSIXct object), neg_MW (negative num) and pos_MW
}
\description{
This method transforms the original retrieved operating reserve calls into a smaller (less variables) data.frame.
}
\details{
---------------------------------------------------------

MAIN FUNCTIONS
}
